// src/app/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  Admin: 'Admin',
  Membro: 'Membro'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const StatusChamado = {
  aberto: 'aberto',
  em_atendimento: 'em_atendimento',
  resolvido: 'resolvido',
  fechado: 'fechado',
  cancelado: 'cancelado'
} as const

export type StatusChamado = (typeof StatusChamado)[keyof typeof StatusChamado]